move the MapDetailsHeader out of the components directory
